+++ PHPCSUtils/Utils/Arrays.php
@@ -102,7 +102,7 @@ final class Arrays
 
         // Is this one of the tokens this function handles ?
         if (isset($tokens[$stackPtr]) === false
-            || isset(Collections::arrayTokensBC()[$tokens[$stackPtr]['code']]) === false
+            || isset(Collections::arrayOpenTokensBC()[$tokens[$stackPtr]['code']]) === false
         ) {
             return false;
         }